splited the 4 pages in different classes (files) still need to fix some stuff here and there
