+++ smartcheckin_/lib/Screens/welcomeScreens/ForgotPswd.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'SignUpScreen.dart';
+import 'HomeScreen.dart';
+
+class ForgotPswd extends StatefulWidget {
+  @override
+  _ForgotPswdState createState() => _ForgotPswdState();
+}
+
+class _ForgotPswdState extends State<ForgotPswd> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
