--- smartcheckin_/lib/Screens/welcomeScreens/sign_in_and_out.dart
@@ -1,1 +1,0 @@
-class SignIn {}
