+++ smartcheckin_/pubspec.yaml
@@ -46,7 +46,9 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-     - assets\mainBackgroundImg.jpeg
+    - images/mainbg.jpeg
+    - images/mainbg.jpg
+    - images/transparentbg.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
